+++ cranberry-commons/src/main/java/io/github/ololx/cranberry/commons/scanner/AbstractCompilationTreeScanner.java
@@ -24,7 +24,6 @@ import javax.lang.model.element.Element;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
-import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.function.Predicate;
 
 /**